requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=h27087fc_1
absl-py=1.4.0=pypi_0
aiohttp=3.8.1=py310h5764c6d_1
aiosignal=1.2.0=pyhd8ed1ab_0
arrow=1.2.3=pyhd8ed1ab_0
arrow-cpp=8.0.0=py310h3098874_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
aws-c-common=0.4.57=he6710b0_1
aws-c-event-stream=0.1.6=h2531618_5
aws-checksums=0.1.9=he6710b0_0
aws-sdk-cpp=1.8.185=hce553d0_0
backcall=0.2.0=pyhd3eb1b0_0
binaryornot=0.4.4=py_1
black=22.10.0=pypi_0
blas=1.0=mkl
boost-cpp=1.78.0=he72f1d9_0
bottleneck=1.3.5=py310ha9d4c09_0
bpemb=0.3.0=py_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotlipy=0.7.0=py310h5764c6d_1004
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.12.7=ha878542_0
catalogue=2.0.8=py310hff52083_0
cd=1.1.0=pypi_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py310h74dc2b5_0
chardet=3.0.4=pypi_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.0.4=py310h06a4308_0
colorama=0.4.5=pyhd8ed1ab_0
cookiecutter=2.1.1=pyh6c4a22f_0
cpuonly=2.0=0
cryptography=35.0.0=py310h685ca39_2
cycler=0.11.0=pyhd8ed1ab_0
cymem=2.0.6=py310hd8f1fbe_3
cython-blis=0.7.7=py310h9102076_0
dataclasses=0.8=pyhc8e2a94_3
datasets=2.4.0=pyhd8ed1ab_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
dill=0.3.5.1=pyhd8ed1ab_0
en-core-web-sm=3.4.0=pypi_0
entrypoints=0.4=py310h06a4308_0
evaluate=0.2.2=pyhd8ed1ab_0
executing=0.8.3=pyhd3eb1b0_0
fasttext=0.9.2=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.6.0=pyhd3eb1b0_0
flatbuffers=23.1.4=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=py310h7f8727e_1
fsspec=2022.8.2=pyhd8ed1ab_0
fugashi=1.2.0=pypi_0
gast=0.4.0=pypi_0
gensim=4.2.0=py310hd8f1fbe_0
gflags=2.2.2=he1b5a44_1004
giflib=5.2.1=h7b6447c_0
glog=0.6.0=h6f12383_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
googletrans=3.0.0=pypi_0
grpc-cpp=1.46.1=h33aed49_0
grpcio=1.51.1=pypi_0
h11=0.9.0=pypi_0
h2=3.2.0=pypi_0
hpack=3.0.0=pypi_0
hstspreload=2022.10.1=pypi_0
httpcore=0.9.1=pypi_0
httpx=0.13.3=pypi_0
huggingface_hub=0.9.1=py_0
hupper=1.10.2=py_0
hyperframe=5.2.0=pypi_0
icu=70.1=h27087fc_0
idna=2.10=pypi_0
importlib-metadata=4.11.3=py310h06a4308_0
importlib_metadata=4.11.3=hd3eb1b0_0
install=1.3.5=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipadic=1.0.0=pypi_0
ipykernel=6.9.1=py310h06a4308_0
ipython=8.4.0=py310h06a4308_0
jedi=0.18.1=py310h06a4308_1
jinja2=3.1.2=pyhd8ed1ab_1
jinja2-time=0.2.0=pyhd8ed1ab_3
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
jupyter-contrib-core=0.4.0=pypi_0
jupyter-contrib-nbextensions=0.5.1=pypi_0
jupyter-highlight-selected-word=0.2.0=pypi_0
jupyter-latex-envs=1.4.6=pypi_0
jupyter-nbextensions-configurator=0.5.0=pypi_0
jupyter_client=7.3.5=py310h06a4308_0
jupyter_core=4.10.0=py310h06a4308_0
keras=2.11.0=pypi_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.2=py310h295c915_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7b6447c_0
langcodes=3.3.0=pyhd8ed1ab_0
language-tool-python=2.7.1=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=12_linux64_mkl
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcblas=3.9.0=12_linux64_mkl
libclang=15.0.6.1=pypi_0
libcurl=7.84.0=h91b91d3_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=h9b69904_4
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=12.1.0=h69a702a_16
libgfortran5=12.1.0=hdcd56e2_16
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=12_linux64_mkl
libnghttp2=1.46.0=hce63b2e_0
libopenblas=0.3.20=pthreads_h78a6416_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.1=h4ff587b_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libthrift=0.15.0=he6d91bd_0
libtiff=4.4.0=hecacb30_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=py310h5764c6d_1
matplotlib-base=3.4.3=py310h23f4a51_2
matplotlib-inline=0.1.6=py310h06a4308_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
multidict=6.0.2=py310h5764c6d_1
multiprocess=0.70.12.2=py310h5764c6d_2
murmurhash=1.0.7=py310h295c915_0
mypy-extensions=0.4.3=pypi_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py310h06a4308_0
nettle=3.7.3=hbbd107a_1
nltk=3.6.7=pyhd8ed1ab_0
numexpr=2.8.3=py310hcea2de6_0
numpy=1.23.1=py310h1794996_0
numpy-base=1.23.1=py310hcba007f_0
oauthlib=3.2.2=pypi_0
openattack=2.1.1=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
opt-einsum=3.3.0=pypi_0
orc=1.7.4=h07ed6aa_0
p2j=1.3.2=pypi_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.3=py310h6a678d5_0
parso=0.8.3=pyhd3eb1b0_0
pastedeploy=2.1.1=pyh9f0ad1d_0
pathspec=0.10.1=pypi_0
pathy=0.6.2=pyhd8ed1ab_0
patsy=0.5.3=pyhd8ed1ab_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py310hace64e9_1
pip=22.1.2=py310h06a4308_0
plaster=1.0=py_0
plaster_pastedeploy=0.7=py_0
platformdirs=2.5.2=pypi_0
preshed=3.0.6=py310h295c915_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.20.1=py310h295c915_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=8.0.0=py310h468efa6_0
pybind11=2.10.0=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pydantic=1.9.0=py310h7f8727e_1
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyramid=2.0=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.4=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-slugify=7.0.0=pyhd8ed1ab_0
python-xxhash=3.0.0=py310h5764c6d_1
python_abi=3.10=2_cp310
pytorch=1.12.1=py3.10_cpu_0
pytorch-crf=0.7.2=pypi_0
pytorch-mutex=1.0=cpu
pytz=2022.2.1=pyhd8ed1ab_0
pyyaml=6.0=py310h5764c6d_4
pyzmq=23.2.0=py310h6a678d5_0
re2=2022.04.01=h27087fc_0
readline=8.1.2=h7f8727e_1
regex=2022.7.9=py310h5eee18b_0
requests=2.28.1=pyhd8ed1ab_1
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pyhd8ed1ab_0
rfc3986=1.5.0=pypi_0
sacremoses=master=py_0
scikit-learn=1.1.1=py310hffb9edd_0
scipy=1.8.1=py310h7612f91_0
seaborn=0.12.1=hd8ed1ab_0
seaborn-base=0.12.1=pyhd8ed1ab_0
sentencepiece=0.1.95=pypi_0
setuptools=63.4.1=py310h06a4308_0
shellingham=1.5.0=pyhd8ed1ab_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
smart_open=5.2.1=pyhd8ed1ab_0
snappy=1.1.9=hbd366e4_1
spacy=3.3.1=py310h3c18c91_0
spacy-legacy=3.0.10=pyhd8ed1ab_0
spacy-loggers=1.0.3=pyhd8ed1ab_0
sqlite=3.39.2=h5082296_0
srsly=2.4.3=py310h295c915_0
stack_data=0.2.0=pyhd3eb1b0_0
statsmodels=0.13.2=py310h96516ba_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.11.0=pypi_0
tensorflow-estimator=2.11.0=pypi_0
tensorflow-hub=0.12.0=pypi_0
tensorflow-io-gcs-filesystem=0.29.0=pypi_0
text-unidecode=1.3=py_0
thinc=8.0.15=py310h2571103_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.11.4=py310h3dcd8bd_1
tomli=2.0.1=pypi_0
torchaudio=0.12.1=py310_cpu
torchvision=0.13.1=py310_cpu
tornado=6.2=py310h5eee18b_0
tqdm=4.64.0=py310h06a4308_0
traitlets=5.1.1=pyhd3eb1b0_0
transformers=4.21.3=py_0
translationstring=1.4=pyh9f0ad1d_0
typer=0.4.2=pyhd8ed1ab_0
typing-extensions=4.3.0=hd8ed1ab_0
typing_extensions=4.3.0=pyha770c72_0
tzdata=2022a=hda174b7_0
unidecode=1.3.6=pyhd8ed1ab_0
urllib3=1.26.11=pyhd8ed1ab_0
utf8proc=2.6.1=h27cfd23_0
venusian=3.0.0=py_0
wasabi=0.10.0=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
webob=1.8.7=pyhd8ed1ab_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
xxhash=0.8.0=h7f98852_3
xz=5.2.5=h7f8727e_1
yaml=0.2.5=h7b6447c_0
yarl=1.7.2=py310h5764c6d_2
zeromq=4.3.4=h2531618_0
zipp=3.8.0=py310h06a4308_0
zlib=1.2.12=h5eee18b_3
zope.deprecation=4.4.0=py_0
zope.interface=5.4.0=py310h5764c6d_2
zstd=1.5.2=ha4553b6_0
